Change adjustment to rounding errors.
In live use, we ran into a NotImplementedError because the rounded
backorders added up to only 33 shirts. Change the algorithm to instead
adjust the sizes in order of worst rounding error.

For example, if (S, M, L) gave divided backorders of (1.38, 2.49, 3.62)
rounding off to (1, 2, 4) or 7 shirts:

- If we wanted 8 shirts, first adjust the 2.49 up to 3 to give (1, 3,
  4).
- If we wanted 9 shirts, then adjust the 1.38 up to 2 to give (2, 3, 4).
- If we wanted 6 shirts, first adjust the 3.62 down to 3 to give (1, 2,
  3).
- If we wanted 5 shirts, then adjust the 1.38 down to 0 to give (0, 2,
  3).

Note this is a materially different algorithm than before.